+++ examples-pydantic-v1/shop_pydantic_v1/client/_api_client.py
@@ -14,7 +14,7 @@ class ShopClient:
     ShopClient
 
     Generated with:
-        pygen = 0.22.1
+        pygen = 0.23.0
         cognite-sdk = 6.33.2
         pydantic = 1.10.7