+++ examples/scenario_instance/client/_api_client.py
@@ -13,7 +13,7 @@ class ScenarioInstanceClient:
     ScenarioInstanceClient
 
     Generated with:
-        pygen = 0.22.1
+        pygen = 0.23.0
         cognite-sdk = 6.33.2
         pydantic = 2.4.2